update DOM with vDom based on socket update
